✨ Enable PDF generation via API GET
