+++ server/controllers/pdf/pdfController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express'
-import { getFooter, getHeader, pdfOptions } from '../../utils/pdfFormat'
+import { configureReportData, getFooter, getHeader, pdfOptions } from '../../utils/pdfFormat'
 import logger from '../../../logger'
 import config from '../../config'
 import ReportService from '../../services/reportService'
@@ -8,23 +8,11 @@ import Report from '../../repositories/entities/report'
 export default class PdfController {
   constructor(protected readonly reportService: ReportService = null) {}
 
-  private configureReportData(report: Report) {
-    const reportData = {
-      reportStatus: report.status,
-      reportType: report.reportDefinition.type,
-      reportVersion: report.reportDefinition.version,
-    }
-    report.fieldValues.forEach(value => {
-      reportData[value.field.name] = value.value
-    })
-    return reportData
-  }
-
   preview = async (req: Request, res: Response): Promise<void> => {
     const { reportId } = req.params
     const report: Report = await this.reportService.getReportById(reportId)
     if (report) {
-      const reportData = { ...this.configureReportData(report), preview: true }
+      const reportData = { ...configureReportData(report), preview: true }
       const headerHtml = getHeader()
       const footerHtml = getFooter({ version: reportData.reportVersion })
       logger.info(`Request to preview ${reportData.reportType} report ${reportId}`)
@@ -40,7 +28,7 @@ export default class PdfController {
     logger.info(`Request to print PDF for report ${reportId}`)
 
     if (report) {
-      const reportData = this.configureReportData(report)
+      const reportData = configureReportData(report)
       const headerHtml = getHeader()
       const footerHtml = getFooter({ version: reportData.reportVersion })
       // Specify preSentenceUrl so that it is used in the NJK template as http://host.docker.internal:3000/assets
