+++ server/routes/api/index.ts
@@ -8,6 +8,7 @@ export default function Index(reportService: ReportService, eventService: EventS
   const apiHandlers = new ApiController(reportService, eventService)
 
   router.get('/v1/report/:id', apiHandlers.getReportById)
+  router.get('/v1/report/:id/pdf', apiHandlers.getPdfById)
   router.get('/v1/reports/:reportType', apiHandlers.getAllReportsByType)
   router.post('/v1/report/:reportType', apiHandlers.createReport)
